+++ console/cypress/integration/events/create-trigger/test.ts
@@ -6,12 +6,13 @@ import { setMetaData } from '../../validators/validators';
 import {
   checkCreateTriggerRoute,
   failCTWithoutData,
-  passCT,
+  passCT2,
   failCTDuplicateTrigger,
   insertTableRow,
   deleteCTTestTrigger,
   deleteCTTestTable,
   passPTCreateTable,
+  passCT1,
   createEtTransform,
   modifyEtTransform,
   deleteEtTransform,
@@ -37,7 +38,8 @@ export const runCreateTriggerTests = () => {
       checkCreateTriggerRoute
     );
     it('Fails to create trigger without data', failCTWithoutData);
-    it('Successfuly creates trigger', passCT);
+    it('Successfuly creates trigger with selected columns for update', passCT1);
+    it('Successfuly creates trigger with all columns for update', passCT2);
     it('Fails to create duplicate trigger', failCTDuplicateTrigger);
     it('Insert a row and invoke trigger', insertTableRow);
     it("Delete's the test trigger", deleteCTTestTrigger);