+++ console/cypress/integration/validators/validators.ts
@@ -475,7 +475,8 @@ export const validateCTrigger = (
   triggerName: string,
   tableName: string,
   schemaName = 'public',
-  result: ResultType
+  result: ResultType,
+  allCols?: boolean
 ) => {
   const reqBody = {
     type: 'export_metadata',
@@ -505,7 +506,11 @@ export const validateCTrigger = (
       expect(response.status === 200).to.be.true;
       expect(trigger.definition.insert.columns === '*').to.be.true;
       expect(trigger.definition.delete.columns === '*').to.be.true;
-      expect(trigger.definition.update.columns.length === 3).to.be.true;
+      if (allCols) {
+        expect(trigger.definition.update.columns === '*').to.be.true;
+      } else {
+        expect(trigger.definition.update.columns.length === 2).to.be.true;
+      }
       expect(
         trigger.retry_conf.interval_sec === parseInt(getIntervalSeconds(), 10)
       ).to.be.true;