+++ console/src/components/Services/Events/EventTriggers/Common/ColumnList.tsx
@@ -2,15 +2,26 @@ import React from 'react';
 import { QualifiedTable } from '@/metadata/types';
 import styles from '../TableCommon/EventTable.scss';
 import { ETOperationColumn } from '../../types';
+import { ColumnSelectionRadioButton } from './ColumnSelectionRadioButton';
 
 type ColumnListProps = {
   operationColumns: ETOperationColumn[];
   table: QualifiedTable;
+  isAllColumnChecked: boolean;
+  readOnlyMode: boolean;
+  handleToggleAllColumn: () => void;
   handleOperationsColumnsChange: (oc: ETOperationColumn[]) => void;
 };
 
 const ColumnList: React.FC<ColumnListProps> = props => {
-  const { operationColumns, table, handleOperationsColumnsChange } = props;
+  const {
+    operationColumns,
+    table,
+    isAllColumnChecked,
+    readOnlyMode,
+    handleToggleAllColumn,
+    handleOperationsColumnsChange,
+  } = props;
 
   if (!table.name) {
     return <i>Select a table first to get column list</i>;
@@ -28,24 +39,37 @@ const ColumnList: React.FC<ColumnListProps> = props => {
 
   return (
     <>
-      {operationColumns.map(opCol => (
-        <div
-          key={opCol.name}
-          className={`${styles.padd_remove} ${styles.columnListElement}`}
-        >
-          <div className="checkbox ">
-            <label className={styles.cursorPointer}>
-              <input
-                type="checkbox"
-                checked={opCol.enabled}
-                onChange={() => handleToggleColumn(opCol)}
-                className={`${styles.cursorPointer} legacy-input-fix`}
-              />
-              {opCol.name}
-            </label>
-          </div>
-        </div>
-      ))}
+      <ColumnSelectionRadioButton
+        isAllColumnChecked={isAllColumnChecked}
+        handleColumnRadioButton={handleToggleAllColumn}
+        readOnly={readOnlyMode}
+      />
+      <div className="mt-sm">
+        {!isAllColumnChecked ? (
+          <>
+            <div>List of columns to select:</div>
+            {operationColumns.map(opCol => (
+              <div
+                key={opCol.name}
+                className={`${styles.padd_remove} ${styles.columnListElement}`}
+              >
+                <div className="checkbox">
+                  <label className="cursor-pointer">
+                    <input
+                      type="checkbox"
+                      checked={opCol.enabled}
+                      onChange={() => handleToggleColumn(opCol)}
+                      className={`${styles.cursorPointer} legacy-input-fix`}
+                      data-test="select-column"
+                    />
+                    {opCol.name}
+                  </label>
+                </div>
+              </div>
+            ))}
+          </>
+        ) : null}
+      </div>
     </>
   );
 };